fix: `Error: Form ... does no exist`: - change subtype value declared in `columnHandler.js` and its occurrencies in `CreateColumn.vue`; - rename the corresponding .vue files; - refactor the corresponding imports.
