+++ src/shared/components/ncTable/mixins/columnHandler.js
@@ -8,8 +8,8 @@ export const ColumnTypes = {
 	Selection: 'selection',
 	SelectionCheck: 'selection-check',
 	SelectionMulti: 'selection-multi',
-	TextIPv4Address: 'text-ipv4address',
-	TextIPv6Address: 'text-ipv6address',
+	TextIPv4Address: 'text-ipv4_address',
+	TextIPv6Address: 'text-ipv6_address',
 	TextLine: 'text-line',
 	TextLink: 'text-link',
 	TextLong: 'text-long',
